# src/main/java/iot/home/service/iotrest/controllers/SensorsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.2.1).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package iot.home.service.iotrest.controllers;

import iot.home.service.iotrest.dto.PostPutReturn;
import iot.home.service.iotrest.dto.Sensor;
import iot.home.service.iotrest.dto.SensorPrediction;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen")
@Validated
@Tag(name = "sensors", description = "the sensors API")
public interface SensorsApi {

    /**
     * DELETE /sensors/{sensor-id}
     * Deletes an existing sensor
     *
     * @param sensorId  (required)
     * @return Sensor deleted successfully (status code 200)
     */
    @Operation(
        operationId = "deleteSensorById",
        tags = { "Sensors" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Sensor deleted successfully")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/sensors/{sensor-id}"
    )
    ResponseEntity<Void> deleteSensorById(
        @Parameter(name = "sensor-id", description = "", required = true) @PathVariable("sensor-id") String sensorId
    );


    /**
     * GET /sensors/home/{home-id}
     * Returns a list of sensors for the specified home
     *
     * @param homeId  (required)
     * @return Returned a list of sensors (status code 200)
     */
    @Operation(
        operationId = "getSensorsByHomeId",
        tags = { "Sensors" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Returned a list of sensors", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Sensor.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/sensors/home/{home-id}",
        produces = { "application/json" }
    )
    ResponseEntity<List<Sensor>> getSensorsByHomeId(
        @Parameter(name = "home-id", description = "", required = true) @PathVariable("home-id") Integer homeId
    );


    /**
     * GET /sensors/{sensor-id}
     * Returns a sensor
     *
     * @param sensorId  (required)
     * @return Returned a sensors (status code 200)
     */
    @Operation(
        operationId = "getSensorsById",
        tags = { "Sensors" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Returned a sensors", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SensorPrediction.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/sensors/{sensor-id}",
        produces = { "application/json" }
    )
    ResponseEntity<List<SensorPrediction>> getSensorsById(
        @Parameter(name = "sensor-id", description = "", required = true) @PathVariable("sensor-id") String sensorId
    );


    /**
     * PUT /sensors
     * Creates new sensor with the specified data.
     *
     * @param sensor  (required)
     * @return Sensor created successfully. (status code 200)
     */
    @Operation(
        operationId = "putSensor",
        tags = { "Sensors" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Sensor created successfully.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = PostPutReturn.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/sensors",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<PostPutReturn> putSensor(
        @Parameter(name = "Sensor", description = "", required = true) @Valid @RequestBody Sensor sensor
    );

}