# src/main/resources/swagger.yml
openapi: 3.0.1
info:
  version: 1.0.11

  title: Climate Web Service
  description: REST api to store data from temperature sensors
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://beta.roger.com/v1
...
components:
  schemas:
...
    Sensor:
      type: object
      properties:
        sensor_id:
          type: string
        home_id:
          type: integer
          format: int32
        room:
          type: string
    SensorPrediction:
      type: object
      properties:
        sensor_id:
          type: string
        home_id:
          type: integer
          format: int32
        room:
          type: string
        temperature:
          type: number
          format: double
...
    PostPutReturn:
      type: object
      properties:
        url:
          type: string
paths:
...
  /sensors:
    put:
      tags:
        - Sensors
      operationId: putSensor
      description:
        Creates new sensor with the specified data.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Sensor'
      responses:
        '200':
          description:
            Sensor created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostPutReturn'
  /sensors/home/{home-id}:
    parameters:
      - name: "home-id"
        in: path
        required: true
        schema:
          type: integer
          format: int32
    get:
      tags:
        - Sensors
      operationId: getSensorsByHomeId
      description:
        Returns a list of sensors for the specified home
      responses:
        '200':
          description:
            Returned a list of sensors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Sensor'
  /sensors/{sensor-id}:
    parameters:
      - name: "sensor-id"
        in: path
        required: true
        schema:
          type: string
    delete:
      tags:
        - Sensors
      operationId: deleteSensorById
      description:
        Deletes an existing sensor
      responses:
        200:
          description:
            Sensor deleted successfully
    get:
      tags:
        - Sensors
      operationId: getSensorsById
      description:
        Returns a sensor
      responses:
        '200':
          description:
            Returned a sensors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SensorPrediction'